ECS Implementation
In this dev session I tried to implement the ECS from simple_ecs and it
worked out great! I am now able to modularly add components to entities
and let the systems handle it all for me.

Pretty straight forward, so I began on implementing player movement
basics with the InputComponent, InputSystem, and the TransformComponent.

Although the player flies around the world and moves, the results are
inconsistent due to the camera.target.

I need to find a better way of handling entity creation and abstracting
from there to better interact with specific objects.

I am referring to my previous code written for Fear The Crow, which
didn't use an ECS,
https://github.com/ionthedev/Fear-the-Crow/tree/acb047adaaae2f83e477ff7d8939735feaa75c58

Build upon this all later